+++ plugins/qtcoinrave/ivselector.cpp
@@ -325,7 +325,7 @@ IvJointDragger::IvJointDragger(QtCoinViewerPtr viewer, ItemPtr pItem, int iSelec
     _jointtype = pjoint->GetType();
     _dofindex = pjoint->GetDOFIndex();
     _jointname = pjoint->GetName();
-    _jointoffset = pjoint->GetOffset();
+    _jointoffset = 0;//pjoint->GetOffset();
     pjoint->GetLimits(_vlower,_vupper);
 
     _pLinkNode = pbody->GetIvLink(iSelectedLink);